Changed the Structure of the programs
