+++ README.md
@@ -1,2 +1,2 @@
 # Database_Server
-Server that handles querying and writing to the database
+tasks.Server that handles querying and writing to the database
