+++ src/main/java/main.java
@@ -1,3 +1,6 @@
+import database.DatabaseManager;
+import tasks.Server;
+
 public class main {
     public static void main(String[] args) {
         DatabaseManager.connect();//start database server
