+++ src/main/java/tasks/Server.java
@@ -1,7 +1,8 @@
+package tasks;
+
 import java.io.*;
 import java.net.ServerSocket;
 import java.net.Socket;
-import java.util.Scanner;
 
 public class Server implements Runnable{
